adding vector add example

+++ Convolution/src/alloc_image_mem/alloc_image_mem.hpp
@@ -2,6 +2,6 @@
 
 #include <inttypes.h>
 
-uint16_t* alloc_image(uint32_t image_size);
-uint16_t* alloc_image_out(uint32_t image_size);
-uint16_t* alloc_kernel(uint32_t kernel_size);
+float* alloc_image(uint32_t image_size);
+float* alloc_image_out(uint32_t image_size);
+float* alloc_kernel(uint32_t kernel_size);
